requirements.txt
click==6.6
Django==1.10.2
django-cors-headers==1.2.2
django-filters==0.2.1
django-markdownx==1.6.3
django-markupfield==1.4.1
djangorestframework==3.4.7
Markdown==2.6.7
Pillow==3.4.1
python-dotenv==0.6.0
gunicorn
